[refactor]: co-locate Wan DiT and architecture config
