+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ exclude: |
     tests/.*|
     scripts/.*|
     fastvideo/dataset/.*|
-    fastvideo/models/.*|
+    fastvideo/models/(?!wan/(config|__init__)\.py$).*|
     ^apps/dreamverse/web/.*|
     examples/.*|
     \.agents/.*|
