+++ docs/design/overview.md
@@ -11,6 +11,8 @@ FastVideo maps a Diffusers-style repo into a pipeline like this:
 - `fastvideo/models/*`: model implementations (DiT, VAE, encoders, upsamplers).
 - `fastvideo/configs/models/*`: arch configs and `param_names_mapping` for
   weight name translation.
+- `fastvideo/models/wan/`: Wan's dense transformer and architecture config
+  together; the old Wan component/config modules remain compatibility imports.
 - `fastvideo/configs/pipelines/*`: pipeline wiring (component classes + names).
 - `fastvideo/api/sampling_param.py`: runtime sampling parameters.
 - `fastvideo/pipelines/basic/*`: end-to-end pipelines.
@@ -48,6 +50,8 @@ runtime parameters consistent:
 
 - `fastvideo/configs/models/`: architecture definitions, layer shapes, and
   `param_names_mapping` rules for key renaming.
+- `fastvideo/models/wan/config.py`: Wan's architecture and mapping rules,
+  co-located with `transformer.py`.
 - `fastvideo/configs/pipelines/`: pipeline wiring and required components.
 - `fastvideo/api/sampling_param.py`: sampling parameters (steps, frames,
   guidance scale, resolution, fps). Defaults come from profiles in
@@ -96,7 +100,8 @@ Note on tensor names:
 
 Official checkpoints often use different `state_dict` names than FastVideo's
 module layout. We translate tensor names via the DiT arch config mapping
-(`param_names_mapping` under `fastvideo/configs/models/dits/`). This is similar
+(`param_names_mapping` under `fastvideo/configs/models/dits/`, or
+`fastvideo/models/wan/config.py` for Wan). This is similar
 in spirit to name-translation layers used in systems like vLLM and SGLang.
 
 Example HF repo (Wan 2.1 T2V 1.3B Diffusers):
@@ -137,7 +142,8 @@ Example `model_index.json` from that repo:
 How this maps to FastVideo:
 
 - `WanPipeline` -> `fastvideo/pipelines/basic/wan/wan_pipeline.py`
-- `WanTransformer3DModel` -> `fastvideo/models/dits/wanvideo.py`
+- `WanTransformer3DModel` -> `fastvideo/models/wan/transformer.py`
+- `WanVideoConfig` -> `fastvideo/models/wan/config.py`
 - `AutoencoderKLWan` -> `fastvideo/models/vaes/wanvae.py`
 - `UMT5EncoderModel` -> `fastvideo/models/encoders/t5.py`
 - `T5TokenizerFast` -> loaded via HF in `fastvideo/models/loader/`
@@ -157,7 +163,7 @@ How this maps to FastVideo:
 
 ## Model components
 
-- DiT models: `fastvideo/models/dits/`
+- DiT models: `fastvideo/models/dits/`; dense Wan: `fastvideo/models/wan/`
 - VAEs: `fastvideo/models/vaes/`
 - Text/image encoders: `fastvideo/models/encoders/`
 - Schedulers: `fastvideo/models/schedulers/`
