+++ fastvideo/models/wan/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Wan components; import config or transformer explicitly to avoid import cycles."""
